File selection
I got rid of edit as I didn't think we needed it. If we do just add it
back. Now able to create a file (with nothing in it of course) as the
file extension of .FHX. Save and load do nothing yet.
