+++ src/prototype1/FHKFilter.java
@@ -1,10 +1,11 @@
+package prototype1;
+
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
 import java.io.File;
-import javax.swing.*;
 import javax.swing.filechooser.*;
 
 /**
